+++ src/map/lua/lua_mobskill.cpp
@@ -68,7 +68,7 @@ inline int32 CLuaMobSkill::setMsg(lua_State *L)
     DSP_DEBUG_BREAK_IF(m_PLuaMobSkill == nullptr);
     DSP_DEBUG_BREAK_IF(lua_isnil(L, -1) || !lua_isnumber(L, -1));
 
-    m_PLuaMobSkill->setMsg(lua_tointeger(L, -1));
+    m_PLuaMobSkill->setMsg((uint16)lua_tointeger(L, -1));
     return 0;
 }
