+++ src/map/lua/luautils.cpp
@@ -1163,7 +1163,7 @@ namespace luautils
             return 0;
         }
 
-        int32 value = lua_tonumber(LuaHandle, -1);
+        int32 value = (int32)lua_tonumber(LuaHandle, -1);
         lua_pop(LuaHandle, -1);
         return value;
     }
@@ -1197,7 +1197,7 @@ namespace luautils
             return 0;
         }
 
-        uint8 value = lua_tonumber(LuaHandle, -1);
+        uint8 value = (uint8)lua_tonumber(LuaHandle, -1);
         lua_pop(LuaHandle, -1);
         return value;
     }
@@ -3096,8 +3096,8 @@ namespace luautils
             lua_pop(LuaHandle, returns);
             return std::tuple<int32, uint8, uint8>();
         }
-        uint8 tpHitsLanded = lua_tonumber(LuaHandle, -4);
-        uint8 extraHitsLanded = lua_tonumber(LuaHandle, -3);
+        uint8 tpHitsLanded = (uint8)lua_tonumber(LuaHandle, -4);
+        uint8 extraHitsLanded = (uint8)lua_tonumber(LuaHandle, -3);
         bool criticalHit = lua_toboolean(LuaHandle, -2);
         int32 dmg = (!lua_isnil(LuaHandle, -1) && lua_isnumber(LuaHandle, -1) ? (int32)lua_tonumber(LuaHandle, -1) : 0);
 
@@ -4272,13 +4272,13 @@ namespace luautils
 
     int32 SetDropRate(lua_State *L)
     {
-        DropList_t* DropList = itemutils::GetDropList(lua_tointeger(L, 1));
+        DropList_t* DropList = itemutils::GetDropList((uint16)lua_tointeger(L, 1));
 
         for (uint8 i = 0; i < DropList->size(); ++i)
         {
             if (DropList->at(i).ItemID == lua_tointeger(L, 2))
             {
-                DropList->at(i).DropRate = lua_tointeger(L, 3);
+                DropList->at(i).DropRate = (uint16)lua_tointeger(L, 3);
                 return 1;
             }
         }
@@ -4327,7 +4327,7 @@ namespace luautils
     {
         if (!lua_isnil(L, 1) && lua_isnumber(L, 1))
         {
-            CAbility* PAbility = ability::GetAbility(lua_tointeger(L, 1));
+            CAbility* PAbility = ability::GetAbility((uint16)lua_tointeger(L, 1));
 
             lua_getglobal(L, CLuaAbility::className);
             lua_pushstring(L, "new");
@@ -4417,16 +4417,16 @@ namespace luautils
 
         position_t center;
         lua_getfield(L, 1, "x");
-        center.x = lua_tonumber(L, -1);
+        center.x = (float)lua_tonumber(L, -1);
         lua_getfield(L, 1, "y");
-        center.y = lua_tonumber(L, -1);
+        center.y = (float)lua_tonumber(L, -1);
         lua_getfield(L, 1, "z");
-        center.z = lua_tonumber(L, -1);
+        center.z = (float)lua_tonumber(L, -1);
         lua_getfield(L, 1, "rot");
-        center.rotation = lua_tonumber(L, -1);
+        center.rotation = (uint8)lua_tonumber(L, -1);
 
-        float radius = lua_tonumber(L, 2);
-        float theta = lua_tonumber(L, 3);
+        float radius = (float)lua_tonumber(L, 2);
+        float theta = (float)lua_tonumber(L, 3);
 
         position_t pos = nearPosition(center, radius, theta);
